+++ waggle-dance-core/src/main/java/com/hotels/bdp/waggledance/mapping/service/impl/StaticDatabaseMappingService.java
@@ -241,9 +241,7 @@ public class StaticDatabaseMappingService implements MappingEventListener {
       log
           .debug("Database Name `{}` maps to metastore with name '{}'", databaseName,
               databaseMapping.getMetastoreMappingName());
-      if (includeInResults(databaseMapping)) {
-        return databaseMapping;
-      }
+      return databaseMapping;
     }
     log.debug("Database Name `{}` not mapped", databaseName);
     throw new NoSuchObjectException("Primary metastore does not have database " + databaseName);