+++ waggle-dance-core/src/test/java/com/hotels/bdp/waggledance/client/ThriftMetastoreClientManagerIntegrationTest.java
@@ -13,21 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-/**
- * Copyright (C) 2016-2023 Expedia, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
 package com.hotels.bdp.waggledance.client;
 
 import static org.hamcrest.CoreMatchers.containsString;
@@ -39,6 +24,7 @@ import static com.hotels.bdp.waggledance.client.HiveUgiArgsStub.TEST_ARGS;
 import org.apache.hadoop.hive.conf.HiveConf;
 import org.apache.hadoop.hive.conf.HiveConf.ConfVars;
 import org.apache.hadoop.hive.metastore.api.Database;
+import org.apache.thrift.TException;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
@@ -88,7 +74,7 @@ public class ThriftMetastoreClientManagerIntegrationTest {
 
     try {
       manager.open(TEST_ARGS);
-    } catch (RuntimeException e) {
+    } catch (TException e) {
       assertThat(e.getMessage(), containsString("java.net.ConnectException: Connection refused"));
     }
   }