+++ waggle-dance-core/src/test/java/com/hotels/bdp/waggledance/client/ThriftMetastoreClientManagerTest.java
@@ -13,21 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-/**
- * Copyright (C) 2016-2023 Expedia, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
 package com.hotels.bdp.waggledance.client;
 
 import static org.mockito.Mockito.never;
@@ -39,6 +24,7 @@ import static com.hotels.bdp.waggledance.client.HiveUgiArgsStub.TEST_ARGS;
 
 import org.apache.hadoop.hive.conf.HiveConf;
 import org.apache.hadoop.hive.conf.HiveConf.ConfVars;
+import org.apache.thrift.TException;
 import org.apache.thrift.transport.TSocket;
 import org.junit.Before;
 import org.junit.Test;
@@ -125,8 +111,8 @@ public class ThriftMetastoreClientManagerTest {
     client.close();
   }
 
-  @Test(expected = RuntimeException.class)
-  public void openSlowConnection() {
+  @Test(expected = TException.class)
+  public void openSlowConnection() throws Exception {
     client = new ThriftMetastoreClientManager(hiveConf, hiveCompatibleThriftHiveMetastoreIfaceFactory, 1);
     client.open(TEST_ARGS);
   }