+++ waggle-dance-core/src/test/java/com/hotels/bdp/waggledance/mapping/model/ASTQueryMappingTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2016-2021 Expedia, Inc.
+ * Copyright (C) 2016-2025 Expedia, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -134,6 +134,14 @@ public class ASTQueryMappingTest {
     query = "SELECT hellobdp() as q union all SELECT db1.hellobdp() as qq where false";
     assertThat(queryMapping.transformOutboundDatabaseName(metaStoreMapping, query),
         is("SELECT hellobdp() as q union all SELECT " + PREFIX + "db1.hellobdp() as qq where false"));
+
+    query = "SELECT COALESCE(db1.hellobdp(`table1.id`, 1)) where false";
+    assertThat(queryMapping.transformOutboundDatabaseName(metaStoreMapping, query),
+        is("SELECT COALESCE(" + PREFIX + "db1.hellobdp(`table1.id`, 1)) where false"));
+
+    query = "SELECT COALESCE(db1.hellobdp(`db2.fun2`(`table1.id`), 1)) where false";
+    assertThat(queryMapping.transformOutboundDatabaseName(metaStoreMapping, query),
+        is("SELECT COALESCE(" + PREFIX + "db1.hellobdp(`" + PREFIX + "db2.fun2`(`table1.id`), 1)) where false"));
   }
 
   @Test