+++ waggle-dance-core/src/test/java/com/hotels/bdp/waggledance/mapping/service/impl/PrefixBasedDatabaseMappingServiceTest.java
@@ -13,21 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-/**
- * Copyright (C) 2016-2023 Expedia, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
 package com.hotels.bdp.waggledance.mapping.service.impl;
 
 import static org.hamcrest.CoreMatchers.is;
@@ -98,7 +83,11 @@ public class PrefixBasedDatabaseMappingServiceTest {
 
   @Before
   public void init() {
-    metaStoreMappingPrimary = mockNewMapping(true, "");
+    init(true, "");
+  }
+
+  private void init(boolean primaryIsAvailable, String primaryPrefix) {
+    metaStoreMappingPrimary = mockNewMapping(primaryIsAvailable, primaryPrefix);
     when(metaStoreMappingPrimary.getClient()).thenReturn(primaryDatabaseClient);
     when(metaStoreMappingPrimary.getLatency()).thenReturn(LATENCY);
     metaStoreMappingFederated = mockNewMapping(true, DB_PREFIX);
@@ -234,40 +223,25 @@ public class PrefixBasedDatabaseMappingServiceTest {
   }
 
   @Test
-  public void databaseMappingMapsToEmptyPrefix() throws NoSuchObjectException {
+  public void databaseMappingMapsToEmptyPrefix() {
     DatabaseMapping databaseMapping = service.databaseMapping("some_unknown_prefix_db");
     assertThat(databaseMapping.getDatabasePrefix(), is(""));
   }
 
   @Test
-  public void databaseMappingDefaultsToPrimaryWhenNothingMatches() throws NoSuchObjectException {
+  public void databaseMappingDefaultsToPrimaryWhenNothingMatches() {
+    String prefix = "primary_";
+    init(true, prefix);
     DatabaseMapping databaseMapping = service.databaseMapping("some_unknown_prefix_db");
-    assertThat(databaseMapping.getDatabasePrefix(), is(""));
+    assertThat(databaseMapping.getDatabasePrefix(), is(prefix));
   }
 
-  @Test(expected = NoSuchObjectException.class)
-  public void databaseMappingDefaultsToPrimaryEvenWhenNothingMatchesAndUnavailable() throws NoSuchObjectException {
-    Mockito.reset(metaStoreMappingPrimary);
-    when(metaStoreMappingPrimary.isAvailable()).thenReturn(false);
+  @Test
+  public void databaseMappingDefaultsToPrimaryEvenWhenNothingMatchesAndUnavailable() {
+    String prefix = "primary_";
+    init(false, prefix);
     DatabaseMapping databaseMapping = service.databaseMapping("some_unknown_prefix_db");
-
-    assertThat(databaseMapping.getDatabasePrefix(), is(""));
-  }
-
-  @Test(expected = NoSuchObjectException.class)
-  public void databaseMappingDoesNotMatchPrimary() throws NoSuchObjectException {
-    AbstractMetaStore noMappedDbsPrimary = primaryMetastore;
-    noMappedDbsPrimary.setMappedDatabases(Collections.emptyList());
-    service.onUpdate(primaryMetastore, noMappedDbsPrimary);
-    service.databaseMapping("some_unknown_db");
-  }
-
-  @Test(expected = NoSuchObjectException.class)
-  public void databaseMappingDoesNotMatchPrimaryWithOtherMappedDbs() throws NoSuchObjectException {
-    AbstractMetaStore noMappedDbsPrimary = primaryMetastore;
-    noMappedDbsPrimary.setMappedDatabases(Collections.singletonList(PRIMARY_DB));
-    service.onUpdate(primaryMetastore, noMappedDbsPrimary);
-    service.databaseMapping("some_unknown_db");
+    assertThat(databaseMapping.getDatabasePrefix(), is(prefix));
   }
 
   @Test
@@ -607,13 +581,12 @@ public class PrefixBasedDatabaseMappingServiceTest {
     federatedMetastore.setMappedDatabases(Collections.singletonList("testName"));
     metaStoreMappingFederated = mockNewMapping(true, DB_PREFIX);
     when(metaStoreMappingFactory.newInstance(federatedMetastore)).thenReturn(metaStoreMappingFederated);
-    when(metaStoreMappingFederated.transformInboundDatabaseName(DB_PREFIX + testDatabase)).thenReturn(testDatabase);
 
     service = new PrefixBasedDatabaseMappingService(metaStoreMappingFactory,
         Arrays.asList(primaryMetastore, federatedMetastore), queryMapping);
 
     DatabaseMapping mapping = service.databaseMapping(DB_PREFIX + testDatabase);
-    assertThat(mapping.getDatabasePrefix(), is(""));
+    assertThat(mapping.getDatabasePrefix(), is(DB_PREFIX));
   }
 
   @Test